pyproject.toml
[project]
name = "fileforge"
version = "0.1.0"
description = "Convert any file into LLM-ready JSON data with chunking support"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Sheraz", email = "sheraz@example.com" }
]
keywords = ["llm", "file-converter", "chunking", "document-processing", "ocr"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # FastAPI and server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.10",
    "asyncpg>=0.30.0",
    # Async tasks
    "celery>=5.4.0",
    "redis>=5.2.0",
    # Data validation
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    # HTTP clients
    "httpx>=0.28.0",
    # Document parsing - Unstructured.io (handles PDF, DOCX, XLSX, HTML, etc.)
    "unstructured[all-docs]>=0.16.0",
    # Additional parsing support
    "pandas>=2.2.0",             # CSV/Excel data handling
    "pillow>=10.0.0",            # Image handling
    # Chunking & Tokenization
    "tiktoken>=0.8.0",           # Token counting (OpenAI tokenizer)
    # Utilities
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.18",  # File uploads
    "python-magic>=0.4.27",      # File type detection
    "aiofiles>=24.1.0",          # Async file operations
]

[dependency-groups]
dev = [
    # Testing
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    # Linting and formatting
    "ruff>=0.8.0",
    # Type checking
    "mypy>=1.13.0",
    "types-redis>=4.6.0",
    # Development tools
    "ipython>=8.31.0",
    "ipdb>=0.13.13",
    "watchdog>=6.0.0",
    # Process management
    "honcho>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["packages", "apps"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=packages --cov=apps --cov-report=term-missing"
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
follow_imports = "normal"
ignore_missing_imports = true

[tool.coverage.run]
source = ["packages", "apps"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
